+++ src/lib.rs
@@ -1,5 +1,6 @@
 #![deny(unsafe_op_in_unsafe_fn)]
 
+pub mod alert_log;
 pub mod attack_events;
 mod attack_timings;
 pub mod config;